Update README.md
The project focuses on analyzing and presenting findings about which parameters influence feeling of 
Happiness at people. The subject itself seems too abstract to be analyzed, but there are methods to quantify 
parameters that influence how people feel.   
While the project has provided analysis and subsequent conclusions the topic is ever changing subject and we 
are learning more and more about it, it is fair to conclude that the project has given several answers but the 
work on understanding Happiness is a never-ending Challenge.